+++ Quiz/Helpers/DeliveryStatus.cs
@@ -1,0 +1,6 @@
+﻿public enum DeliveryStatus : byte
+{
+    Delivered,
+    InProgress,
+    NotDelivered
+}
