+++ Quiz/Models/OrderDetails.cs
@@ -1,0 +1,9 @@
+﻿namespace Quiz.models
+{
+    public class OrderDetails
+    {
+        public int Id { get; set; }
+        public string? CustomerName { get; set; }
+        public string? ContactNumber { get; set; }
+    }
+}
