+++ Quiz/Services/Email.cs
@@ -2,9 +2,9 @@
 {
     public class Email
     {
-        public string From { get; set; }
-        public string To { get; set; }
-        public string Subject { get; set; }
-        public string Body { get; set; }
+        public string? From { get; set; }
+        public string? To { get; set; }
+        public string? Subject { get; set; }
+        public string? Body { get; set; }
     }
 }
